+++ src/main/java/org/jabref/gui/preferences/customentrytypes/CustomEntryTypesTab.java
@@ -1,12 +1,10 @@
-package org.jabref.gui.customentrytypes;
+package org.jabref.gui.preferences.customentrytypes;
 
 import javafx.application.Platform;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.fxml.FXML;
 import javafx.scene.Group;
 import javafx.scene.control.Button;
-import javafx.scene.control.ButtonBar.ButtonData;
-import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
@@ -20,14 +18,14 @@ import javafx.scene.input.Dragboard;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.input.TransferMode;
 
-import org.jabref.gui.DialogService;
 import org.jabref.gui.DragAndDropDataFormats;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.icon.IconTheme;
-import org.jabref.gui.theme.ThemeManager;
-import org.jabref.gui.util.BaseDialog;
+import org.jabref.gui.preferences.AbstractPreferenceTabView;
+import org.jabref.gui.preferences.PreferencesTab;
 import org.jabref.gui.util.ControlHelper;
 import org.jabref.gui.util.CustomLocalDragboard;
+import org.jabref.gui.util.FieldsUtil;
 import org.jabref.gui.util.ValueTableCellFactory;
 import org.jabref.gui.util.ViewModelTableRowFactory;
 import org.jabref.logic.l10n.Localization;
@@ -35,17 +33,13 @@ import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.entry.field.Field;
-import org.jabref.preferences.PreferencesService;
 
 import com.airhacks.afterburner.views.ViewLoader;
 import com.tobiasdiez.easybind.EasyBind;
 import de.saxsys.mvvmfx.utils.validation.visualization.ControlsFxVisualizer;
 import jakarta.inject.Inject;
 
-public class CustomizeEntryTypeDialogView extends BaseDialog<Void> {
-
-    private final BibDatabaseMode mode;
-    private final BibEntryTypesManager entryTypesManager;
+public class CustomEntryTypesTab extends AbstractPreferenceTabView<CustomEntryTypesTabViewModel> implements PreferencesTab {
 
     @FXML private TableView<EntryTypeViewModel> entryTypesTable;
     @FXML private TableColumn<EntryTypeViewModel, String> entryTypColumn;
@@ -57,47 +51,36 @@ public class CustomizeEntryTypeDialogView extends BaseDialog<Void> {
     @FXML private TableColumn<FieldViewModel, String> fieldTypeActionColumn;
     @FXML private TableColumn<FieldViewModel, Boolean> fieldTypeMultilineColumn;
     @FXML private ComboBox<Field> addNewField;
-    @FXML private ButtonType applyButton;
-    @FXML private ButtonType resetButton;
     @FXML private Button addNewEntryTypeButton;
     @FXML private Button addNewFieldButton;
 
-    @Inject private PreferencesService preferencesService;
-
     @Inject private StateManager stateManager;
-    @Inject private DialogService dialogService;
-    @Inject private ThemeManager themeManager;
 
-    private CustomEntryTypeDialogViewModel viewModel;
     private final ControlsFxVisualizer visualizer = new ControlsFxVisualizer();
+
     private CustomLocalDragboard localDragboard;
 
-    public CustomizeEntryTypeDialogView(BibDatabaseContext bibDatabaseContext, BibEntryTypesManager entryTypesManager) {
-        this.setTitle(Localization.lang("Customize entry types"));
-        this.mode = bibDatabaseContext.getMode();
-        this.entryTypesManager = entryTypesManager;
-
+    public CustomEntryTypesTab() {
         ViewLoader.view(this)
-                  .load()
-                  .setAsDialogPane(this);
-
-        setResultConverter(button -> {
-            if (button.getButtonData() == ButtonData.OK_DONE) {
-                viewModel.apply();
-            }
-            return null;
-        });
-        ControlHelper.setAction(resetButton, getDialogPane(), event -> this.resetEntryTypes());
-
-        themeManager.updateFontStyle(getDialogPane().getScene());
+                  .root(this)
+                  .load();
     }
 
-    @FXML
-    private void initialize() {
+    @Override
+    public String getTabName() {
+        return Localization.lang("Entry types");
+    }
+
+    public void initialize() {
+        BibDatabaseMode mode = stateManager.getActiveDatabase().map(BibDatabaseContext::getMode)
+                                           .orElse(preferencesService.getGeneralPreferences().getDefaultBibDatabaseMode());
+        BibEntryTypesManager entryTypesRepository = preferencesService.getCustomEntryTypesRepository();
+
+        this.viewModel = new CustomEntryTypesTabViewModel(mode, entryTypesRepository, dialogService, preferencesService);
+
         // As the state manager gets injected it's not available in the constructor
         this.localDragboard = stateManager.getLocalDragboard();
 
-        viewModel = new CustomEntryTypeDialogViewModel(mode, preferencesService, entryTypesManager, dialogService);
         setupEntryTypesTable();
         setupFieldsTable();
 
@@ -108,15 +91,6 @@ public class CustomizeEntryTypeDialogView extends BaseDialog<Void> {
             visualizer.initVisualization(viewModel.entryTypeValidationStatus(), addNewEntryType, true);
             visualizer.initVisualization(viewModel.fieldValidationStatus(), addNewField, true);
         });
-    }
-
-    private void makeRotatedColumnHeader(TableColumn<?, ?> column, String text) {
-        Label label = new Label();
-        label.setText(text);
-        label.setRotate(-90);
-        label.setMinWidth(80);
-        column.setGraphic(new Group(label));
-        column.getStyleClass().add("rotated");
     }
 
     private void setupEntryTypesTable() {
@@ -161,32 +135,32 @@ public class CustomizeEntryTypeDialogView extends BaseDialog<Void> {
             if (type != null) {
                 var items = type.fields();
                 fields.setItems(items);
+            } else {
+                fields.setItems(null);
             }
         });
     }
 
     private void setupFieldsTable() {
-        fieldNameColumn.setCellValueFactory(item -> item.getValue().fieldName());
+        fieldNameColumn.setCellValueFactory(item -> item.getValue().nameProperty());
 
         fieldTypeColumn.setCellFactory(CheckBoxTableCell.forTableColumn(fieldTypeColumn));
-        fieldTypeColumn.setCellValueFactory(item -> item.getValue().fieldTypeRequired());
+        fieldTypeColumn.setCellValueFactory(item -> item.getValue().requiredProperty());
         makeRotatedColumnHeader(fieldTypeColumn, Localization.lang("Required"));
 
         fieldTypeMultilineColumn.setCellFactory(CheckBoxTableCell.forTableColumn(fieldTypeMultilineColumn));
-        fieldTypeMultilineColumn.setCellValueFactory(item -> item.getValue().multiline());
+        fieldTypeMultilineColumn.setCellValueFactory(item -> item.getValue().multilineProperty());
         makeRotatedColumnHeader(fieldTypeMultilineColumn, Localization.lang("Multiline"));
 
         fieldTypeActionColumn.setSortable(false);
         fieldTypeActionColumn.setReorderable(false);
         fieldTypeActionColumn.setEditable(false);
-        fieldTypeActionColumn.setCellValueFactory(cellData -> cellData.getValue().fieldName());
+        fieldTypeActionColumn.setCellValueFactory(cellData -> cellData.getValue().nameProperty());
 
         new ValueTableCellFactory<FieldViewModel, String>()
                 .withGraphic(item -> IconTheme.JabRefIcons.DELETE_ENTRY.getGraphicNode())
                 .withTooltip(name -> Localization.lang("Remove field %0 from currently selected entry type", name))
-                .withOnMouseClickedEvent(item -> evt -> {
-                    viewModel.removeField(fields.getSelectionModel().getSelectedItem());
-                })
+                .withOnMouseClickedEvent(item -> evt -> viewModel.removeField(fields.getSelectionModel().getSelectedItem()))
                 .install(fieldTypeActionColumn);
 
         new ViewModelTableRowFactory<FieldViewModel>()
@@ -197,22 +171,33 @@ public class CustomizeEntryTypeDialogView extends BaseDialog<Void> {
                 .install(fields);
 
         addNewField.setItems(viewModel.fieldsForAdding());
-        addNewField.setConverter(CustomEntryTypeDialogViewModel.FIELD_STRING_CONVERTER);
+        addNewField.setConverter(FieldsUtil.FIELD_STRING_CONVERTER);
 
         viewModel.newFieldToAddProperty().bindBidirectional(addNewField.valueProperty());
         // The valueProperty() of addNewField ComboBox needs to be updated by typing text in the ComboBox textfield,
         // since the enabled/disabled state of addNewFieldButton won't update otherwise
-        EasyBind.subscribe(addNewField.getEditor().textProperty(), text -> addNewField.setValue(CustomEntryTypeDialogViewModel.FIELD_STRING_CONVERTER.fromString(text)));
+        EasyBind.subscribe(addNewField.getEditor().textProperty(), text -> addNewField.setValue(FieldsUtil.FIELD_STRING_CONVERTER.fromString(text)));
     }
 
-    private void handleOnDragOver(TableRow<FieldViewModel> row, FieldViewModel originalItem, DragEvent event) {
+    private void makeRotatedColumnHeader(TableColumn<?, ?> column, String text) {
+        Label label = new Label();
+        label.setText(text);
+        label.setRotate(-90);
+        label.setMinWidth(80);
+        column.setGraphic(new Group(label));
+        column.getStyleClass().add("rotated");
+    }
+
+    private void handleOnDragOver(TableRow<FieldViewModel> row, FieldViewModel originalItem, DragEvent
+            event) {
         if ((event.getGestureSource() != originalItem) && event.getDragboard().hasContent(DragAndDropDataFormats.FIELD)) {
             event.acceptTransferModes(TransferMode.MOVE);
             ControlHelper.setDroppingPseudoClasses(row, event);
         }
     }
 
-    private void handleOnDragDetected(TableRow<FieldViewModel> row, FieldViewModel fieldViewModel, MouseEvent event) {
+    private void handleOnDragDetected(TableRow<FieldViewModel> row, FieldViewModel fieldViewModel, MouseEvent
+            event) {
         row.startFullDrag();
         FieldViewModel field = fields.getSelectionModel().getSelectedItem();
 
@@ -257,15 +242,18 @@ public class CustomizeEntryTypeDialogView extends BaseDialog<Void> {
         viewModel.addNewField();
     }
 
-    private void resetEntryTypes() {
+    @FXML
+    void resetEntryTypes() {
         boolean reset = dialogService.showConfirmationDialogAndWait(
-                                            Localization.lang("Reset entry types and fields to defaults"),
-                                            Localization.lang("This will reset all entry types to their default values and remove all custom entry types"),
-                                            Localization.lang("Reset to default"));
+                Localization.lang("Reset entry types and fields to defaults"),
+                Localization.lang("This will reset all entry types to their default values and remove all custom entry types"),
+                Localization.lang("Reset to default"));
         if (reset) {
             viewModel.resetAllCustomEntryTypes();
-            viewModel.addAllTypes();
-            this.entryTypesTable.refresh();
+            fields.getSelectionModel().clearSelection();
+            entryTypesTable.getSelectionModel().clearSelection();
+            viewModel.setValues();
+            entryTypesTable.refresh();
         }
     }
 }